requirements.txt
openai
python-dotenv
fastapi
pydantic